requirements.txt
streamlit
plotly.express
pandas
matplotlib
BeautifulSoup4
requests